abstract context away from middleware for audio switching
